refactor: streamline source architecture
