+++ scripts/check-file-size.mjs
@@ -3,8 +3,6 @@ import path from "node:path";
 
 const DEFAULT_MAX_LINES = 500;
 const LINE_BUDGETS = new Map([
-  ["src/file-store.ts", 585],
-  ["src/permissions.ts", 566],
   ["src/root-impl.ts", 1750],
   ["src/root-path.ts", 862],
   ["test/api-coverage.test.ts", 915],
