--- src/file-url.ts
@@ -1,1 +1,0 @@
-export * from "./local-file-access.js";
