+++ src/short-path.ts
@@ -1,0 +1,6 @@
+import os from "node:os";
+
+export function shortPath(value: string): string {
+  const home = os.homedir();
+  return value.startsWith(home) ? `~${value.slice(home.length)}` : value;
+}
